Increase conflicting task chunks day orders if day order is provided
This commit changes the semantics of changing the day order of a task
chunk. Previously, it was completely ignored if the day was
changed, and otherwise the day order exchanged with the chunk that
currently held that day order.

This commit increases all day orders (i.e., moves the chunks down) if
the provided day order is already in use by a chunk on the day to which
the chunk is moved. This includes both updates that do not changed the
day and updates moving between different days.